update polymer fixtures.
