+++ hoomd_polymers/tests/base/test_molecule.py
@@ -191,58 +191,36 @@ class TestPolymer(BaseTest):
 
 
 class TestCopolymer(BaseTest):
-    class COC(Polymer):
-        def __init__(self, lengths, num_mols, **kwargs):
-            smiles = BaseTest.dimethylether_smiles
-            bond_indices = [3, -1]
-            bond_length = 0.15
-            bond_orientation = [None, None]
-            super().__init__(
-                lengths=lengths, num_mols=num_mols,
-                smiles=smiles, bond_indices=bond_indices,
-                bond_length=bond_length, bond_orientation=bond_orientation,
-                **kwargs)
-
-    class CC(Polymer):
-        def __init__(self, lengths, num_mols, **kwargs):
-            smiles = BaseTest.ethane_smiles
-            bond_indices = [2, -2]
-            bond_length = 0.15
-            bond_orientation = [None, None]
-            super().__init__(
-                lengths=lengths, num_mols=num_mols,
-                smiles=smiles, bond_indices=bond_indices,
-                bond_length=bond_length, bond_orientation=bond_orientation,
-                **kwargs)
-
-    def test_copolymer_with_sequence(self):
-        copolymer = CoPolymer(monomer_A=TestCopolymer.COC,
-                              monomer_B=TestCopolymer.CC,
+    def test_copolymer_with_sequence(self, PolyEthylene, PolyDME):
+        copolymer = CoPolymer(monomer_A=PolyDME,
+                              monomer_B=PolyEthylene,
                               lengths=1, num_mols=1, sequence="ABA")
         assert copolymer.n_particles == 22
         assert ('C', 'C', 'C', 'C') in copolymer.topology_information[
             "dihedral_types"]
 
-    def test_copolymer_with_sequence_different_chain_lengths(self):
-        copolymer = CoPolymer(monomer_A=TestCopolymer.COC,
-                              monomer_B=TestCopolymer.CC,
+    def test_copolymer_with_sequence_different_chain_lengths(self, PolyEthylene,
+                                                             PolyDME):
+        copolymer = CoPolymer(monomer_A=PolyDME,
+                              monomer_B=PolyEthylene,
                               lengths=[2, 3], num_mols=[1, 1], sequence="ABA")
 
         assert copolymer.molecules[0].n_particles == 42
         assert copolymer.molecules[1].n_particles == 62
 
-    def test_copolymer_with_sequence_different_num_mol(self):
-        copolymer = CoPolymer(monomer_A=TestCopolymer.COC,
-                              monomer_B=TestCopolymer.CC,
+    def test_copolymer_with_sequence_different_num_mol(self, PolyEthylene,
+                                                       PolyDME):
+        copolymer = CoPolymer(monomer_A=PolyDME,
+                              monomer_B=PolyEthylene,
                               lengths=[2, 3], num_mols=[1, 2], sequence="ABA")
 
         assert copolymer.molecules[0].n_particles == 42
         assert copolymer.molecules[1].n_particles == 62
         assert copolymer.molecules[2].n_particles == 62
 
-    def test_copolymer_random_sequence(self):
-        copolymer = CoPolymer(monomer_A=TestCopolymer.COC,
-                              monomer_B=TestCopolymer.CC,
+    def test_copolymer_random_sequence(self, PolyEthylene, PolyDME):
+        copolymer = CoPolymer(monomer_A=PolyDME,
+                              monomer_B=PolyEthylene,
                               lengths=[3], num_mols=[1], random_sequence=True,
                               seed=42)
         # sequence is BAA
